se resuelve error de checkFechaAndSync cuando no habia miembros

+++ src/main/java/com/roshka/dtaporteria/service/Scheduler.java
@@ -112,24 +112,26 @@ public class Scheduler {
     public void checkFechaAndSync(){
         List<MemberDTO>members = memberService.list();
         List<Member>resultMembers = new LinkedList<>();
-        Map<String, MemberDTO> fecha_vencimiento = members.stream().collect(Collectors.toMap(MemberDTO::getId, Function.identity()));
-        fecha_vencimiento.forEach(
-                (key, value) -> {
-                    if (!Objects.equals(value.getType(), "Socio") && !isAfterCurrentDate(value.getFecha_vencimiento())){
-                        try {
-                            memberService.delete(key);
-                        } catch (Exception e) {
-                            Date now = new Date();
-                            String fecha = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.ENGLISH).format(now);
-                            logErroresRepository.save(new LogErrores(fecha,
-                                    "No se pudo borrar un miembro con fecha de vencimiento invalida"));
+        if (members != null) {
+            Map<String, MemberDTO> fecha_vencimiento = members.stream().collect(Collectors.toMap(MemberDTO::getId, Function.identity()));
+            fecha_vencimiento.forEach(
+                    (key, value) -> {
+                        if (!Objects.equals(value.getType(), "Socio") && !isAfterCurrentDate(value.getFecha_vencimiento())){
+                            try {
+                                memberService.delete(key);
+                            } catch (Exception e) {
+                                Date now = new Date();
+                                String fecha = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss", Locale.ENGLISH).format(now);
+                                logErroresRepository.save(new LogErrores(fecha,
+                                        "No se pudo borrar un miembro con fecha de vencimiento invalida"));
+                            }
+                        }
+                        else{
+                            resultMembers.add(membersPg(value));
                         }
                     }
-                    else{
-                        resultMembers.add(membersPg(value));
-                    }
-                }
-        );
+            );
+        }
         membersRepository.saveAll(resultMembers);
     }
